refactor(#1459): Extract TupleRepository from ReBACManager god object (Phase 7)
Extract data access layer into tuples/repository.py (~1021 LOC):
- Connection management (get_connection, close_connection, connection ctx mgr)
- SQL dialect abstraction (fix_sql_placeholders, supports_old_new_returning)
- Zone revision tracking (get_zone_revision, increment_zone_revision)
- Pure tuple queries (find_subject_sets, find_related_objects, etc.)
- Cycle detection (recursive CTE)
- Cross-zone validation
- ABAC condition evaluation

ReBACManager composes TupleRepository via self._repo and delegates
all extracted methods. Reduces rebac_manager.py from 4,599 to 3,826 LOC
(-773 LOC, -17%).

All 48 core ReBAC tests pass. No behavioral changes.
